Tidy up React logic, pulled from #2096
