+++ src/utils/json.js
@@ -12,7 +12,7 @@ type JSONValue = Array<JSONValue> | string | number | JSON;
 type JSON = { [key: string]: JSONValue };
 
 // this will mutate the original JSON object
-export function mergeAdacentJSONTextNodes(node: JSON, removeFunctions: boolean, visitedNodes?: Set<JSON>) {
+export function mergeAdjacentJSONTextNodes(node: JSON, removeFunctions: boolean, visitedNodes?: Set<JSON>) {
   if (visitedNodes === undefined) {
     visitedNodes = new Set();
   }
@@ -41,7 +41,7 @@ export function mergeAdacentJSONTextNodes(node: JSON, removeFunctions: boolean, 
           arr.push(concatString);
           concatString = null;
         }
-        arr.push(mergeAdacentJSONTextNodes(child, removeFunctions, visitedNodes));
+        arr.push(mergeAdjacentJSONTextNodes(child, removeFunctions, visitedNodes));
       }
     }
     if (concatString !== null) {
@@ -58,7 +58,7 @@ export function mergeAdacentJSONTextNodes(node: JSON, removeFunctions: boolean, 
           node[key] = "function";
         }
       } else if (typeof value === "object" && value !== null) {
-        node[key] = mergeAdacentJSONTextNodes(((value: any): JSON), removeFunctions, visitedNodes);
+        node[key] = mergeAdjacentJSONTextNodes(((value: any): JSON), removeFunctions, visitedNodes);
       }
     }
   }
